Introduce EndpointConfig factoring common TCP/QUIC endpoint configuration.
