+++ vertx-core/src/main/java/io/vertx/core/net/QuicEndpointConfig.java
@@ -21,51 +21,35 @@ import java.util.Objects;
  * @author <a href="mailto:julien@julienviet.com">Julien Viet</a>
  */
 @DataObject
-public abstract class QuicEndpointConfig {
+public abstract class QuicEndpointConfig extends EndpointConfig {
 
-  private QuicConfig transportConfig;
   private QLogConfig qlogConfig;
   private String keyLogFile;
-  private Duration idleTimeout;
-  private Duration readIdleTimeout;
-  private Duration writeIdleTimeout;
-  private String metricsName;
-  private NetworkLogging networkLogging;
 
   public QuicEndpointConfig() {
-    this.transportConfig = new QuicConfig();
+    super();
+    setTransportConfig(new QuicConfig());
     this.qlogConfig = null;
-    this.idleTimeout = null;
-    this.readIdleTimeout = null;
-    this.writeIdleTimeout = null;
-    this.metricsName = null;
-    this.networkLogging = null;
   }
 
   public QuicEndpointConfig(QuicEndpointConfig other) {
+    super(other);
 
     QLogConfig qLogConfig = other.qlogConfig;
 
-    this.transportConfig = other.transportConfig.copy();
     this.qlogConfig = qLogConfig != null ? new QLogConfig(qLogConfig) : null;
     this.keyLogFile = other.keyLogFile;
-    this.idleTimeout = other.idleTimeout;
-    this.readIdleTimeout = other.readIdleTimeout;
-    this.writeIdleTimeout = other.writeIdleTimeout;
-    this.metricsName = other.metricsName;
-    this.networkLogging = other.networkLogging != null ? new NetworkLogging(other.networkLogging) : null;
   }
 
   /**
    * @return the endpoint transport config
    */
   public QuicConfig getTransportConfig() {
-    return transportConfig;
+    return (QuicConfig)super.getTransportConfig();
   }
 
   public QuicEndpointConfig setTransportConfig(QuicConfig transportConfig) {
-    this.transportConfig = Objects.requireNonNull(transportConfig);
-    return this;
+    return (QuicEndpointConfig)super.setTransportConfig(transportConfig);
   }
 
   /**
@@ -113,107 +97,23 @@ public abstract class QuicEndpointConfig {
     return this;
   }
 
-  /**
-   * Set the stream idle timeout, zero or {@code null} means don't time out.
-   * This determines if a stream will timeout and be closed if no data is received nor sent within the timeout.
-   *
-   * @param idleTimeout  the idle timeout
-   * @return a reference to this, so the API can be used fluently
-   */
   public QuicEndpointConfig setIdleTimeout(Duration idleTimeout) {
-    if (idleTimeout != null && idleTimeout.isNegative()) {
-      throw new IllegalArgumentException("streamIdleTimeout must be >= 0");
-    }
-    this.idleTimeout = idleTimeout;
-    return this;
+    return (QuicEndpointConfig) super.setIdleTimeout(idleTimeout);
   }
 
-  /**
-   * @return the idle timeout applied to each stream
-   */
-  public Duration getIdleTimeout() {
-    return idleTimeout;
-  }
-
-  /**
-   * <p>Set the stream read idle timeout, zero or {@code null} means or null means don't time out. This determines if a
-   * stream will timeout and be closed if no data is received within the timeout.</p>
-   *
-   * @param idleTimeout  the read idle timeout
-   * @return a reference to this, so the API can be used fluently
-   */
   public QuicEndpointConfig setReadIdleTimeout(Duration idleTimeout) {
-    if (idleTimeout != null && idleTimeout.isNegative()) {
-      throw new IllegalArgumentException("streamReadIdleTimeout must be >= 0");
-    }
-    this.readIdleTimeout = idleTimeout;
-    return this;
+    return (QuicEndpointConfig) super.setReadIdleTimeout(idleTimeout);
   }
 
-  /**
-   * @return the read idle timeout applied to each stream
-   */
-  public Duration getReadIdleTimeout() {
-    return readIdleTimeout;
-  }
-
-  /**
-   * <p>Set the stream write idle timeout, zero or {@code null} means don't time out. This determines if a
-   * stream will timeout and be closed if no data is sent within the timeout.</p>
-   *
-   * @param idleTimeout  the write idle timeout
-   * @return a reference to this, so the API can be used fluently
-   */
   public QuicEndpointConfig setWriteIdleTimeout(Duration idleTimeout) {
-    if (idleTimeout != null && idleTimeout.isNegative()) {
-      throw new IllegalArgumentException("streamWriteIdleTimeout must be >= 0");
-    }
-    this.writeIdleTimeout = idleTimeout;
-    return this;
+    return (QuicEndpointConfig) super.setWriteIdleTimeout(idleTimeout);
   }
 
-  /**
-   * @return the write idle timeout applied to each stream
-   */
-  public Duration getWriteIdleTimeout() {
-    return writeIdleTimeout;
-  }
-
-
-  /**
-   * @return the metrics name identifying the reported metrics.
-   */
-  public String getMetricsName() {
-    return metricsName;
-  }
-
-  /**
-   * Set the metrics name identifying the reported metrics, useful for grouping metrics
-   * with the same name.
-   *
-   * @param metricsName the metrics name
-   * @return a reference to this, so the API can be used fluently
-   */
   public QuicEndpointConfig setMetricsName(String metricsName) {
-    this.metricsName = metricsName;
-    return this;
+    return (QuicEndpointConfig) super.setMetricsName(metricsName);
   }
 
-  /**
-   * @return the stream network logging config, {@code null} means disabled
-   */
-  public NetworkLogging getNetworkLogging() {
-    return networkLogging;
-  }
-
-  /**
-   * Configure the per stream networking logging: Netty's stream pipeline is configured for logging on Netty's logger.
-   *
-   * @param config the stream network logging config, {@code null} means disabled
-   * @return a reference to this, so the API can be used fluently
-   */
   public QuicEndpointConfig setNetworkLogging(NetworkLogging config) {
-    this.networkLogging = config;
-    return this;
+    return (QuicEndpointConfig) super.setNetworkLogging(config);
   }
 }
